requirements.txt
fastapi
uvicorn
tortoise-orm
aerich
httpx
pytest
pytest-asyncio